@@ -1,6 +1,6 @@
 {
   "name": "wikibase-solutions/php-cypher-dsl",
-  "description": "A query builder for the Cypher query language written in PHP",
+  "description": "A query builder for the Cypher query language",
   "type": "library",
   "keywords": [
     "neo4j",
@@ -46,18 +46,15 @@
     "source": "https://github.com/WikibaseSolutions/php-cypher-dsl"
   },
   "require": {
-    "php": ">=7.4",
-    "ext-ctype": "*",
-    "ext-openssl": "*",
-    "symfony/polyfill-php80": "^1.25",
-    "symfony/polyfill-php81": "^1.25"
+    "php": ">=8.1",
+    "ext-ctype": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "~9.0",
-    "infection/infection": "^0.25.5",
+    "phpunit/phpunit": "^10.0",
+    "infection/infection": "^0.29",
     "friendsofphp/php-cs-fixer": "^3.0",
     "rregeer/phpunit-coverage-check": "^0.3.1",
-    "phpstan/phpstan": "^1.8"
+    "phpstan/phpstan": "^2.0"
   },
   "autoload": {
     "files": [
