@@ -12,9 +12,8 @@ namespace WikibaseSolutions\CypherDSL\Clauses;
 use WikibaseSolutions\CypherDSL\Patterns\Pattern;
 use WikibaseSolutions\CypherDSL\QueryConvertible;
 use WikibaseSolutions\CypherDSL\Syntax\Alias;
-use WikibaseSolutions\CypherDSL\Traits\CastTrait;
-use WikibaseSolutions\CypherDSL\Traits\ErrorTrait;
 use WikibaseSolutions\CypherDSL\Types\AnyType;
+use WikibaseSolutions\CypherDSL\Utils\CastUtils;
 
 /**
  * This class represents a RETURN clause.
@@ -25,34 +24,29 @@ use WikibaseSolutions\CypherDSL\Types\AnyType;
  */
 final class ReturnClause extends Clause
 {
-    use CastTrait;
-    use ErrorTrait;
-
     /**
      * @var bool Whether to be a RETURN DISTINCT query
      */
     private bool $distinct = false;
 
     /**
-     * @var Alias[]|AnyType[]|(Alias|AnyType)[] The expressions to return
+     * @var (Alias|AnyType)[] The expressions to return
      */
     private array $columns = [];
 
     /**
      * Add a new column to this RETURN clause.
      *
-     * @param Alias|AnyType|bool|float|int|mixed[]|Pattern|string ...$columns The values to return
-     *
-     * @return $this
+     * @param Alias|AnyType|array|bool|float|int|Pattern|string ...$columns The values to return
      *
      * @see https://neo4j.com/docs/cypher-manual/current/clauses/return/#return-column-alias
      */
-    public function addColumn(...$columns): self
+    public function addColumn(Alias|AnyType|bool|float|int|array|Pattern|string ...$columns): self
     {
         $res = [];
 
         foreach ($columns as $column) {
-            $res[] = $column instanceof Alias ? $column : self::toAnyType($column);
+            $res[] = $column instanceof Alias ? $column : CastUtils::toAnyType($column);
         }
 
         $this->columns = array_merge($this->columns, $res);
@@ -62,8 +56,6 @@ final class ReturnClause extends Clause
 
     /**
      * Sets this query to only return unique rows.
-     *
-     * @return $this
      *
      * @see https://neo4j.com/docs/cypher-manual/current/clauses/return/#return-unique-results
      */
@@ -77,7 +69,7 @@ final class ReturnClause extends Clause
     /**
      * Returns the columns to return. Aliased columns have string keys instead of integers.
      *
-     * @return Alias[]|AnyType[]|(Alias|AnyType)[]
+     * @return (Alias|AnyType)[]
      */
     public function getColumns(): array
     {
