@@ -9,14 +9,16 @@
  */
 namespace WikibaseSolutions\CypherDSL\Expressions\Literals;
 
-use WikibaseSolutions\CypherDSL\Traits\CastTrait;
-use WikibaseSolutions\CypherDSL\Traits\EscapeTrait;
+use Stringable;
+use WikibaseSolutions\CypherDSL\Patterns\Pattern;
 use WikibaseSolutions\CypherDSL\Traits\TypeTraits\CompositeTypeTraits\MapTypeTrait;
 use WikibaseSolutions\CypherDSL\Types\AnyType;
 use WikibaseSolutions\CypherDSL\Types\CompositeTypes\MapType;
+use WikibaseSolutions\CypherDSL\Utils\CastUtils;
+use WikibaseSolutions\CypherDSL\Utils\NameUtils;
 
 /**
- * This class represents a CYPHER map. For example, this class can represent the following
+ * This class represents a Cypher map. For example, this class can represent the following
  * construct:.
  *
  * {name: 'Andy', sport: 'Brazilian Ju-Jitsu'}
@@ -26,8 +28,6 @@ use WikibaseSolutions\CypherDSL\Types\CompositeTypes\MapType;
  */
 final class Map implements MapType
 {
-    use CastTrait;
-    use EscapeTrait;
     use MapTypeTrait;
 
     /**
@@ -36,26 +36,26 @@ final class Map implements MapType
     private array $elements;
 
     /**
-     * @param mixed[] $elements Associative array of the elements that this map should have
+     * @param array $elements Associative array of the elements that this map should have
      *
      * @internal This method is not covered by the backwards compatibility promise of php-cypher-dsl
      */
     public function __construct(array $elements = [])
     {
-        $this->elements = array_map([self::class, 'toAnyType'], $elements);
+        $this->elements = array_map(CastUtils::toAnyType(...), $elements);
     }
 
     /**
      * Adds an element for the given name with the given value. Overrides the element if the $key already exists.
      *
-     * @param string $key   The name/label for the element
-     * @param mixed  $value The value of the element
+     * @param string                                                 $key   The name/label for the element
+     * @param AnyType|array|bool|float|int|Pattern|string|Stringable $value The value of the element
      *
      * @return $this
      */
-    public function add(string $key, $value): self
+    public function add(string $key, AnyType|Pattern|Stringable|bool|float|int|array|string $value): self
     {
-        $this->elements[$key] = self::toAnyType($value);
+        $this->elements[$key] = CastUtils::toAnyType($value);
 
         return $this;
     }
@@ -100,7 +100,7 @@ final class Map implements MapType
         $pairs = [];
 
         foreach ($this->elements as $key => $value) {
-            $pairs[] = sprintf("%s: %s", self::escape((string) $key), $value->toQuery());
+            $pairs[] = sprintf("%s: %s", NameUtils::escape((string) $key), $value->toQuery());
         }
 
         return sprintf("{%s}", implode(", ", $pairs));
