@@ -11,9 +11,8 @@ namespace WikibaseSolutions\CypherDSL\Patterns;
 
 use WikibaseSolutions\CypherDSL\Expressions\Label;
 use WikibaseSolutions\CypherDSL\Expressions\Literals\Map;
-use WikibaseSolutions\CypherDSL\Traits\ErrorTrait;
-use WikibaseSolutions\CypherDSL\Traits\EscapeTrait;
 use WikibaseSolutions\CypherDSL\Traits\PatternTraits\PropertyPatternTrait;
+use WikibaseSolutions\CypherDSL\Utils\NameUtils;
 
 /**
  * This class represents a node.
@@ -22,9 +21,6 @@ use WikibaseSolutions\CypherDSL\Traits\PatternTraits\PropertyPatternTrait;
  */
 final class Node implements CompletePattern, PropertyPattern, RelatablePattern
 {
-    use ErrorTrait;
-    use EscapeTrait;
-
     use PropertyPatternTrait;
 
     /**
@@ -60,8 +56,6 @@ final class Node implements CompletePattern, PropertyPattern, RelatablePattern
 
     /**
      * Adds one or more labels to this node.
-     *
-     * @param string ...$label
      *
      * @return $this
      */
@@ -146,7 +140,7 @@ final class Node implements CompletePattern, PropertyPattern, RelatablePattern
 
         if ($this->labels !== []) {
             foreach ($this->labels as $label) {
-                $nodeInner .= ":" . self::escape($label);
+                $nodeInner .= ":" . NameUtils::escape($label);
             }
         }
 
