@@ -10,7 +10,6 @@
 namespace WikibaseSolutions\CypherDSL\Patterns;
 
 use WikibaseSolutions\CypherDSL\Expressions\Variable;
-use WikibaseSolutions\CypherDSL\Traits\ErrorTrait;
 use WikibaseSolutions\CypherDSL\Traits\PatternTraits\PatternTrait;
 use WikibaseSolutions\CypherDSL\Traits\TypeTraits\PropertyTypeTraits\BooleanTypeTrait;
 use WikibaseSolutions\CypherDSL\Types\CompositeTypes\MapType;
@@ -24,8 +23,6 @@ use WikibaseSolutions\CypherDSL\Types\PropertyTypes\BooleanType;
 final class Path implements BooleanType, CompletePattern, RelatablePattern
 {
     use BooleanTypeTrait;
-    use ErrorTrait;
-
     use PatternTrait;
 
     /**
@@ -44,13 +41,10 @@ final class Path implements BooleanType, CompletePattern, RelatablePattern
      *
      * @internal This method is not covered by the backwards compatibility guarantee of php-cypher-dsl
      */
-    public function __construct($nodes = [], $relationships = [])
+    public function __construct(Node|array $nodes = [], Relationship|array $relationships = [])
     {
         $nodes = is_array($nodes) ? $nodes : [$nodes];
         $relationships = is_array($relationships) ? $relationships : [$relationships];
-
-        self::assertClassArray('nodes', Node::class, $nodes);
-        self::assertClassArray('relationships', Relationship::class, $relationships);
 
         $this->nodes = $nodes;
         $this->relationships = $relationships;
@@ -100,10 +94,10 @@ final class Path implements BooleanType, CompletePattern, RelatablePattern
     /**
      * @inheritDoc
      */
-    public function relationshipTo(Pattern $pattern, ?string $type = null, $properties = null, $name = null): self
+    public function relationshipTo(Pattern $pattern, ?string $type = null, MapType|array|null $properties = null, Variable|string|null $name = null): self
     {
         return $this->relationship(
-            self::buildRelationship(Relationship::DIR_RIGHT, $type, $properties, $name),
+            self::buildRelationship(Direction::RIGHT, $type, $properties, $name),
             $pattern
         );
     }
@@ -111,10 +105,10 @@ final class Path implements BooleanType, CompletePattern, RelatablePattern
     /**
      * @inheritDoc
      */
-    public function relationshipFrom(Pattern $pattern, ?string $type = null, $properties = null, $name = null): self
+    public function relationshipFrom(Pattern $pattern, ?string $type = null, MapType|array|null $properties = null, Variable|string|null $name = null): self
     {
         return $this->relationship(
-            self::buildRelationship(Relationship::DIR_LEFT, $type, $properties, $name),
+            self::buildRelationship(Direction::LEFT, $type, $properties, $name),
             $pattern
         );
     }
@@ -122,10 +116,10 @@ final class Path implements BooleanType, CompletePattern, RelatablePattern
     /**
      * @inheritDoc
      */
-    public function relationshipUni(Pattern $pattern, ?string $type = null, $properties = null, $name = null): self
+    public function relationshipUni(Pattern $pattern, ?string $type = null, MapType|array|null $properties = null, Variable|string|null $name = null): self
     {
         return $this->relationship(
-            self::buildRelationship(Relationship::DIR_UNI, $type, $properties, $name),
+            self::buildRelationship(Direction::UNI, $type, $properties, $name),
             $pattern
         );
     }
@@ -179,12 +173,12 @@ final class Path implements BooleanType, CompletePattern, RelatablePattern
     /**
      * Construct a new relationship from the given parameters.
      *
-     * @param string[]             $direction  The direction of the relationship (should be a Relationship::DIR_* constant)
+     * @param Direction            $direction  The direction of the relationship
      * @param null|string          $type       The type of the relationship
-     * @param null|MapType|mixed[] $properties The properties to add to the relationship
+     * @param null|array|MapType   $properties The properties to add to the relationship
      * @param null|string|Variable $name       The name of the variable to which to assign this relationship
      */
-    private static function buildRelationship(array $direction, ?string $type = null, $properties = null, $name = null): Relationship
+    private static function buildRelationship(Direction $direction, ?string $type = null, MapType|array|null $properties = null, Variable|string|null $name = null): Relationship
     {
         $relationship = new Relationship($direction);
 
