@@ -43,11 +43,9 @@ interface Pattern extends QueryConvertible
     /**
      * Explicitly assign a named variable to this pattern.
      *
-     * @param null|string|Variable $variable
-     *
      * @return $this
      */
-    public function withVariable($variable): self;
+    public function withVariable(Variable|string|null $variable): self;
 
     /**
      * Returns the variable of this pattern. This function generates a variable if none has been set. It will implicitly set the variable of the pattern as well.
