@@ -24,29 +24,21 @@ interface StringType extends PropertyType
 {
     /**
      * Check whether this expression the given expression.
-     *
-     * @param string|StringType $right
      */
-    public function contains($right, bool $insertParentheses = true): Contains;
+    public function contains(self|string $right, bool $insertParentheses = true): Contains;
 
     /**
      * Perform a suffix string search with the given expression.
-     *
-     * @param string|StringType $right
      */
-    public function endsWith($right, bool $insertParentheses = true): EndsWith;
+    public function endsWith(self|string $right, bool $insertParentheses = true): EndsWith;
 
     /**
      * Perform a prefix string search with the given expression.
-     *
-     * @param string|StringType $right
      */
-    public function startsWith($right, bool $insertParentheses = true): StartsWith;
+    public function startsWith(self|string $right, bool $insertParentheses = true): StartsWith;
 
     /**
      * Perform a regex comparison with the given expression.
-     *
-     * @param string|StringType $right
      */
-    public function regex($right, bool $insertParentheses = true): Regex;
+    public function regex(self|string $right, bool $insertParentheses = true): Regex;
 }
