@@ -7,8 +7,9 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace WikibaseSolutions\CypherDSL\Traits;
+namespace WikibaseSolutions\CypherDSL\Utils;
 
+use Stringable;
 use WikibaseSolutions\CypherDSL\Expressions\Literals\Literal;
 use WikibaseSolutions\CypherDSL\Expressions\Variable;
 use WikibaseSolutions\CypherDSL\Patterns\Pattern;
@@ -23,123 +24,89 @@ use WikibaseSolutions\CypherDSL\Types\PropertyTypes\StringType;
 use WikibaseSolutions\CypherDSL\Types\StructuralTypes\StructuralType;
 
 /**
- * Helper trait for casting native PHP types to Cypher-DSL types. Casts are added to this class on an as-needed basis.
+ * Utility functions for casting native PHP types to Cypher-DSL types. Casts are added to this class on an as-needed
+ * basis.
  *
- * @internal This trait is not covered by the backwards compatibility guarantee of php-cypher-dsl
+ * @internal This class is not covered by the backwards compatibility guarantee of php-cypher-dsl
  */
-trait CastTrait
+class CastUtils
 {
-    use ErrorTrait;
-
     /**
      * Casts the given value to a ListType.
      *
-     * @param ListType|mixed[] $list
+     * @param ListType|ListType[] $list
      */
-    private static function toListType($list): ListType
+    public static function toListType(ListType|array $list): ListType
     {
-        self::assertClass('list', [ListType::class, 'array'], $list);
-
         return $list instanceof ListType ? $list : Literal::list($list);
     }
 
     /**
      * Casts the given value to a MapType.
      *
-     * @param MapType|mixed[] $map
+     * @param MapType|MapType[] $map
      */
-    private static function toMapType($map): MapType
+    public static function toMapType(MapType|array $map): MapType
     {
-        self::assertClass('map', [MapType::class, 'array'], $map);
-
         return $map instanceof MapType ? $map : Literal::map($map);
     }
 
     /**
      * Casts the given value to a StringType.
-     *
-     * @param string|StringType $string
      */
-    private static function toStringType($string): StringType
+    public static function toStringType(StringType|string $string): StringType
     {
-        self::assertClass('string', [StringType::class, 'string'], $string);
-
         return $string instanceof StringType ? $string : Literal::string($string);
     }
 
     /**
      * Casts the given value to a NumeralType.
-     *
-     * @param float|int|NumeralType $numeral
      */
-    private static function toNumeralType($numeral): NumeralType
+    public static function toNumeralType(NumeralType|float|int $numeral): NumeralType
     {
-        self::assertClass('numeral', [NumeralType::class, 'int', 'float'], $numeral);
-
-        // @phpstan-ignore-next-line
         return $numeral instanceof NumeralType ? $numeral : Literal::number($numeral);
     }
 
     /**
      * Casts the given value to an IntegerType.
-     *
-     * @param int|IntegerType $integer
      */
-    private static function toIntegerType($integer): IntegerType
+    public static function toIntegerType(IntegerType|int $integer): IntegerType
     {
-        self::assertClass('integer', [IntegerType::class, 'int'], $integer);
-
         return $integer instanceof IntegerType ? $integer : Literal::integer($integer);
     }
 
     /**
      * Casts the given value to a BooleanType.
-     *
-     * @param bool|BooleanType $boolean
      */
-    private static function toBooleanType($boolean): BooleanType
+    public static function toBooleanType(BooleanType|bool $boolean): BooleanType
     {
-        self::assertClass('boolean', [BooleanType::class, 'bool'], $boolean);
-
         return $boolean instanceof BooleanType ? $boolean : Literal::boolean($boolean);
     }
 
     /**
      * Casts the given value to a PropertyType.
-     *
-     * @param bool|float|int|PropertyType|string $property
      */
-    private static function toPropertyType($property): PropertyType
+    public static function toPropertyType(PropertyType|bool|int|float|string $property): PropertyType
     {
-        self::assertClass('property', [PropertyType::class, 'bool', 'int', 'float', 'string'], $property);
-
         // @phpstan-ignore-next-line
         return $property instanceof PropertyType ? $property : Literal::literal($property);
     }
 
     /**
      * Casts the given value to a StructuralType.
-     *
-     * @param Pattern|StructuralType $structure
      */
-    private static function toStructuralType($structure): StructuralType
+    public static function toStructuralType(StructuralType|Pattern $structure): StructuralType
     {
-        self::assertClass('structure', [Pattern::class, StructuralType::class], $structure);
-
         return $structure instanceof StructuralType ? $structure : $structure->getVariable();
     }
 
     /**
      * Casts the given value to a Variable.
      *
-     * @param Pattern|string|Variable $variable
-     *
-     * @see CastTrait::toName() for a function that does not accept Pattern
+     * @see CastUtils::toName() for a function that does not accept Pattern
      */
-    private static function toVariable($variable): Variable
+    public static function toVariable(Variable|Pattern|string $variable): Variable
     {
-        self::assertClass('variable', [Variable::class, Pattern::class, 'string'], $variable);
-
         if ($variable instanceof Variable) {
             return $variable;
         }
@@ -152,28 +119,20 @@ trait CastTrait
     }
 
     /**
-     * Casts the given value to a name (as a variable).
+     * Casts the given value to a name.
      *
-     * @param string|Variable $name
-     *
-     * @see CastTrait::toVariable() for a function that accepts Pattern
+     * @see CastUtils::toVariable() for a function that accepts Pattern
      */
-    private static function toName($name): Variable
+    public static function toName(Variable|string $name): Variable
     {
-        self::assertClass('name', [Variable::class, 'string'], $name);
-
         return $name instanceof Variable ? $name : new Variable($name);
     }
 
     /**
      * Casts the given value to an AnyType.
-     *
-     * @param AnyType|bool|float|int|mixed[]|Pattern|string $value
      */
-    private static function toAnyType($value): AnyType
+    public static function toAnyType(AnyType|Pattern|Stringable|bool|float|int|array|string $value): AnyType
     {
-        self::assertClass('value', [AnyType::class, Pattern::class, 'int', 'float', 'string', 'bool', 'array'], $value);
-
         if ($value instanceof Pattern) {
             return $value->getVariable();
         }
@@ -182,7 +141,6 @@ trait CastTrait
             return $value;
         }
 
-        // @phpstan-ignore-next-line
         return Literal::literal($value);
     }
 }
