@@ -12,7 +12,7 @@ namespace WikibaseSolutions\CypherDSL\Tests\Unit\Clauses;
 use PHPUnit\Framework\TestCase;
 use TypeError;
 use WikibaseSolutions\CypherDSL\Clauses\MatchClause;
-use WikibaseSolutions\CypherDSL\Patterns\Relationship;
+use WikibaseSolutions\CypherDSL\Patterns\Direction;
 use WikibaseSolutions\CypherDSL\Query;
 
 /**
@@ -90,7 +90,7 @@ final class MatchClauseTest extends TestCase
 
     public function testAddPatternDoesNotAcceptRelationship(): void
     {
-        $rel = Query::relationship(Relationship::DIR_LEFT);
+        $rel = Query::relationship(Direction::LEFT);
 
         $match = new MatchClause();
 
