@@ -10,12 +10,12 @@
 namespace WikibaseSolutions\CypherDSL\Tests\Unit\Patterns;
 
 use DomainException;
-use InvalidArgumentException;
 use LogicException;
 use PHPUnit\Framework\TestCase;
 use WikibaseSolutions\CypherDSL\Expressions\Literals\Integer;
 use WikibaseSolutions\CypherDSL\Expressions\Literals\String_;
 use WikibaseSolutions\CypherDSL\Expressions\Variable;
+use WikibaseSolutions\CypherDSL\Patterns\Direction;
 use WikibaseSolutions\CypherDSL\Patterns\Relationship;
 use WikibaseSolutions\CypherDSL\Query;
 
@@ -24,29 +24,140 @@ use WikibaseSolutions\CypherDSL\Query;
  */
 final class RelationshipTest extends TestCase
 {
-    public function testExceptionIsThrownWhenInvalidDirection(): void
+    public static function provideVariableLengthRelationshipsWithNameData(): array
     {
-        $this->expectException(InvalidArgumentException::class);
-        $r = new Relationship(['--', '--']);
+        return [
+            ['b', 0, 100, Direction::UNI, '-[b*0..100]-'],
+            ['b', 5, 5, Direction::RIGHT, '-[b*5..5]->'],
+            ['a', 10, null, Direction::UNI, '-[a*10..]-'],
+            ['a', null, 10, Direction::LEFT, '<-[a*..10]-'],
+        ];
+    }
+
+    public static function provideVariableLengthRelationshipsWithTypeData(): array
+    {
+        return [
+            ['', 1, 100, Direction::LEFT, '<-[*1..100]-'],
+            ['a', 10, null, Direction::LEFT, '<-[:a*10..]-'],
+            [':', null, 10, Direction::LEFT, '<-[:`:`*..10]-'],
+        ];
+    }
+
+    public static function provideVariableLengthRelationshipsWithPropertiesData(): array
+    {
+        return [
+            [[], 10, 100, Direction::LEFT, "<-[*10..100]-"],
+            [[new String_('a')], 10, null, Direction::LEFT, "<-[*10.. {`0`: 'a'}]-"],
+            [['a' => new String_('b')], null, 10, Direction::LEFT, "<-[*..10 {a: 'b'}]-"],
+        ];
+    }
+
+    public static function provideVariableLengthRelationshipsWithNameAndTypeAndPropertiesData(): array
+    {
+        return [
+            ['a', 'a', [], 10, 100, Direction::LEFT, "<-[a:a*10..100]-"],
+            ['b', 'a', [new String_('a')], null, 10, Direction::LEFT, "<-[b:a*..10 {`0`: 'a'}]-"],
+            ['a', 'b', [new String_('a')], 10, 100, Direction::LEFT, "<-[a:b*10..100 {`0`: 'a'}]-"],
+            ['a', '', ['a' => new String_('b')], null, 10, Direction::LEFT, "<-[a*..10 {a: 'b'}]-"],
+            ['a', ':', ['a' => new String_('b'), new String_('c')], 10, null, Direction::LEFT, "<-[a:`:`*10.. {a: 'b', `0`: 'c'}]-"],
+        ];
+    }
+
+    public static function provideWithNameData(): array
+    {
+        return [
+            ['a', Direction::UNI, '-[a]-'],
+            ['a', Direction::LEFT, '<-[a]-'],
+            ['a', Direction::RIGHT, '-[a]->'],
+        ];
+    }
+
+    public static function provideAddTypeData(): array
+    {
+        return [
+            ['', Direction::LEFT, '<--'],
+            ['a', Direction::LEFT, '<-[:a]-'],
+            [':', Direction::LEFT, '<-[:`:`]-'],
+        ];
+    }
+
+    public static function provideWithTypesData(): array
+    {
+        return [
+            [['', 'a'], Direction::LEFT, '<-[:a]-'],
+            [['a', 'b'], Direction::LEFT, '<-[:a|b]-'],
+            [['', 'a', 'b'], Direction::LEFT, '<-[:a|b]-'],
+            [['a', '', 'b'], Direction::LEFT, '<-[:a|b]-'],
+            [['a', 'b', ':'], Direction::LEFT, '<-[:a|b|`:`]-'],
+        ];
+    }
+
+    public static function provideWithPropertiesData(): array
+    {
+        return [
+            [[], Direction::LEFT, "<--"],
+            [[new String_('a')], Direction::LEFT, "<-[{`0`: 'a'}]-"],
+            [['a' => new String_('b')], Direction::LEFT, "<-[{a: 'b'}]-"],
+            [['a' => new String_('b'), new String_('c')], Direction::LEFT, "<-[{a: 'b', `0`: 'c'}]-"],
+            [[':' => new Integer(12)], Direction::LEFT, "<-[{`:`: 12}]-"],
+            [['a' => 'b', 'c' => 12, 'd' => 12.38], Direction::LEFT, "<-[{a: 'b', c: 12, d: 12.38}]-"],
+        ];
+    }
+
+    public static function provideWithNameAndTypeData(): array
+    {
+        return [
+            ['a', '', Direction::LEFT, '<-[a]-'],
+            ['a', 'b', Direction::LEFT, '<-[a:b]-'],
+        ];
+    }
+
+    public static function provideWithNameAndPropertiesData(): array
+    {
+        return [
+            ['a', [], Direction::LEFT, "<-[a]-"],
+            ['b', [new String_('a')], Direction::LEFT, "<-[b {`0`: 'a'}]-"],
+        ];
+    }
+
+    public static function provideWithNameAndTypeAndPropertiesData(): array
+    {
+        return [
+            ['a', 'a', [], Direction::LEFT, "<-[a:a]-"],
+            ['b', 'a', [new String_('a')], Direction::LEFT, "<-[b:a {`0`: 'a'}]-"],
+            ['a', 'b', [new String_('a')], Direction::LEFT, "<-[a:b {`0`: 'a'}]-"],
+            ['a', '', ['a' => new String_('b')], Direction::LEFT, "<-[a {a: 'b'}]-"],
+            ['a', ':', ['a' => new String_('b'), new String_('c')], Direction::LEFT, "<-[a:`:` {a: 'b', `0`: 'c'}]-"],
+        ];
+    }
+
+    public static function provideWithMultipleTypesData(): array
+    {
+        return [
+            ['a', [], [], Direction::LEFT, "<-[a]-"],
+            ['b', ['a'], [new String_('a')], Direction::LEFT, "<-[b:a {`0`: 'a'}]-"],
+            ['a', ['a', 'b', 'c'], [new String_('a')], Direction::LEFT, "<-[a:a|b|c {`0`: 'a'}]-"],
+            ['a', ['a', 'b'], [], Direction::LEFT, "<-[a:a|b]-"],
+        ];
     }
 
     public function testDirRight(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
 
         $this->assertSame("-->", $r->toQuery());
     }
 
     public function testDirLeft(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
 
         $this->assertSame("<--", $r->toQuery());
     }
 
     public function testDirUni(): void
     {
-        $r = new Relationship(Relationship::DIR_UNI);
+        $r = new Relationship(Direction::UNI);
 
         $this->assertSame("--", $r->toQuery());
     }
@@ -54,7 +165,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideWithNameData
      */
-    public function testWithName(string $name, array $direction, string $expected): void
+    public function testWithName(string $name, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withVariable($name);
@@ -65,7 +176,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideAddTypeData
      */
-    public function testAddType(string $type, array $direction, string $expected): void
+    public function testAddType(string $type, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->addType($type);
@@ -75,7 +186,7 @@ final class RelationshipTest extends TestCase
 
     public function testAddTypeMultiple(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
         $r->addType("a");
         $r->addType("b");
         $r->addType(":");
@@ -86,7 +197,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideWithTypesData
      */
-    public function testWithTypes(array $types, array $direction, string $expected): void
+    public function testWithTypes(array $types, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withTypes($types);
@@ -97,7 +208,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideWithPropertiesData
      */
-    public function testWithProperties(array $properties, array $direction, string $expected): void
+    public function testWithProperties(array $properties, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withProperties($properties);
@@ -108,7 +219,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideWithNameAndTypeData
      */
-    public function testWithNameAndType(string $name, string $type, array $direction, string $expected): void
+    public function testWithNameAndType(string $name, string $type, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withVariable($name)->addType($type);
@@ -118,7 +229,7 @@ final class RelationshipTest extends TestCase
 
     public function testWithNameAndMultipleTypes(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
         $r->withVariable('a')->addType('a')->addType('b');
 
         $this->assertSame('<-[a:a|b]-', $r->toQuery());
@@ -126,7 +237,7 @@ final class RelationshipTest extends TestCase
 
     public function testWithVariableActualVariable(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
         $r->withVariable(new Variable('a'));
 
         $this->assertSame('<-[a]-', $r->toQuery());
@@ -135,7 +246,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideWithNameAndPropertiesData
      */
-    public function testWithNameAndProperties(string $name, array $properties, array $direction, string $expected): void
+    public function testWithNameAndProperties(string $name, array $properties, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withVariable($name)->withProperties($properties);
@@ -146,7 +257,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideWithNameAndTypeAndPropertiesData
      */
-    public function testWithNameAndTypeAndProperties(string $name, string $type, array $properties, array $direction, string $expected): void
+    public function testWithNameAndTypeAndProperties(string $name, string $type, array $properties, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withVariable($name)->addType($type)->withProperties($properties);
@@ -157,7 +268,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideWithMultipleTypesData
      */
-    public function testWithMultipleTypes(string $name, array $types, array $properties, array $direction, string $expected): void
+    public function testWithMultipleTypes(string $name, array $types, array $properties, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withVariable($name)->withProperties($properties);
@@ -172,7 +283,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideVariableLengthRelationshipsWithNameData
      */
-    public function testVariableLengthRelationshipsWithName(string $name, ?int $minHops, ?int $maxHops, array $direction, string $expected): void
+    public function testVariableLengthRelationshipsWithName(string $name, ?int $minHops, ?int $maxHops, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withVariable($name);
@@ -191,7 +302,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideVariableLengthRelationshipsWithTypeData
      */
-    public function testVariableLengthRelationshipsWithType(string $type, ?int $minHops, ?int $maxHops, array $direction, string $expected): void
+    public function testVariableLengthRelationshipsWithType(string $type, ?int $minHops, ?int $maxHops, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->addType($type);
@@ -210,7 +321,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideVariableLengthRelationshipsWithPropertiesData
      */
-    public function testVariableLengthRelationshipsWithProperties(array $properties, ?int $minHops, ?int $maxHops, array $direction, string $expected): void
+    public function testVariableLengthRelationshipsWithProperties(array $properties, ?int $minHops, ?int $maxHops, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withProperties($properties);
@@ -229,7 +340,7 @@ final class RelationshipTest extends TestCase
     /**
      * @dataProvider provideVariableLengthRelationshipsWithNameAndTypeAndPropertiesData
      */
-    public function testVariableLengthRelationshipsWithNameAndTypeAndProperties(string $name, string $type, array $properties, ?int $minHops, ?int $maxHops, array $direction, string $expected): void
+    public function testVariableLengthRelationshipsWithNameAndTypeAndProperties(string $name, string $type, array $properties, ?int $minHops, ?int $maxHops, Direction $direction, string $expected): void
     {
         $r = new Relationship($direction);
         $r->withVariable($name)->addType($type)->withProperties($properties);
@@ -247,7 +358,7 @@ final class RelationshipTest extends TestCase
 
     public function testArbitraryHops(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
         $r->withVariable('hello')->addType('world')->addType('testing')->withProperties(['is' => 'a virtue']);
         $r->withArbitraryHops();
 
@@ -256,7 +367,7 @@ final class RelationshipTest extends TestCase
 
     public function testExactLengthRelationships(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withVariable("tom")
             ->addType("Person")
             ->withProperties(['name' => 'Tom Hanks']);
@@ -268,7 +379,7 @@ final class RelationshipTest extends TestCase
 
     public function testMinAndExactHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withMinHops(1);
 
         $this->expectException(LogicException::class);
@@ -278,7 +389,7 @@ final class RelationshipTest extends TestCase
 
     public function testMaxAndExactHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withMaxHops(1);
 
         $this->expectException(LogicException::class);
@@ -288,7 +399,7 @@ final class RelationshipTest extends TestCase
 
     public function testMinMaxAndExactHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withMinHops(1);
         $r->withMaxHops(1);
 
@@ -299,7 +410,7 @@ final class RelationshipTest extends TestCase
 
     public function testExactAndMinHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withExactHops(1);
 
         $this->expectException(LogicException::class);
@@ -309,7 +420,7 @@ final class RelationshipTest extends TestCase
 
     public function testExactAndMaxHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withExactHops(1);
 
         $this->expectException(LogicException::class);
@@ -319,7 +430,7 @@ final class RelationshipTest extends TestCase
 
     public function testMaxHopsLessThanMinHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withMinHops(100);
 
         $this->expectException(DomainException::class);
@@ -329,7 +440,7 @@ final class RelationshipTest extends TestCase
 
     public function testMinHopsGreaterThanMaxHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withMaxHops(1);
 
         $this->expectException(DomainException::class);
@@ -339,7 +450,7 @@ final class RelationshipTest extends TestCase
 
     public function testMinHopsLessThanZero(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
 
         $this->expectException(DomainException::class);
 
@@ -348,7 +459,7 @@ final class RelationshipTest extends TestCase
 
     public function testMaxHopsLessThanOne(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
 
         $this->expectException(DomainException::class);
 
@@ -357,7 +468,7 @@ final class RelationshipTest extends TestCase
 
     public function testMaxHopsLessThanZero(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
 
         $this->expectException(DomainException::class);
 
@@ -366,7 +477,7 @@ final class RelationshipTest extends TestCase
 
     public function testExactHopsLessThanOne(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
 
         $this->expectException(DomainException::class);
 
@@ -375,7 +486,7 @@ final class RelationshipTest extends TestCase
 
     public function testExactHopsLessThanZero(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
 
         $this->expectException(DomainException::class);
 
@@ -384,7 +495,7 @@ final class RelationshipTest extends TestCase
 
     public function testExactHopsWithArbitraryHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withArbitraryHops();
 
         $this->expectException(LogicException::class);
@@ -394,7 +505,7 @@ final class RelationshipTest extends TestCase
 
     public function testMinHopsWithArbitraryHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withArbitraryHops();
 
         $this->expectException(LogicException::class);
@@ -404,7 +515,7 @@ final class RelationshipTest extends TestCase
 
     public function testMaxHopsWithArbitraryHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withArbitraryHops();
 
         $this->expectException(LogicException::class);
@@ -414,7 +525,7 @@ final class RelationshipTest extends TestCase
 
     public function testArbitraryHopsWithExactHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withExactHops(5);
 
         $this->expectException(LogicException::class);
@@ -424,7 +535,7 @@ final class RelationshipTest extends TestCase
 
     public function testArbitraryHopsWithMinHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withMinHops(5);
 
         $this->expectException(LogicException::class);
@@ -434,7 +545,7 @@ final class RelationshipTest extends TestCase
 
     public function testArbitraryHopsWithMaxHops(): void
     {
-        $r = new Relationship(Relationship::DIR_RIGHT);
+        $r = new Relationship(Direction::RIGHT);
         $r->withMaxHops(5);
 
         $this->expectException(LogicException::class);
@@ -444,14 +555,14 @@ final class RelationshipTest extends TestCase
 
     public function testGetDirection(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
-        $this->assertSame(Relationship::DIR_LEFT, $r->getDirection());
+        $r = new Relationship(Direction::LEFT);
+        $this->assertSame(Direction::LEFT, $r->getDirection());
     }
 
     public function testGetProperties(): void
     {
         $properties = Query::map(['foo' => 'bar']);
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
         $this->assertNull($r->getProperties());
 
         $r->withProperties($properties);
@@ -461,7 +572,7 @@ final class RelationshipTest extends TestCase
 
     public function testGetExactHops(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
         $this->assertNull($r->getExactHops());
 
         $r->withExactHops(12);
@@ -471,7 +582,7 @@ final class RelationshipTest extends TestCase
 
     public function testGetMaxHops(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
         $this->assertNull($r->getMaxHops());
 
         $r->withMaxHops(12);
@@ -481,7 +592,7 @@ final class RelationshipTest extends TestCase
 
     public function testGetMinHops(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
         $this->assertNull($r->getMinHops());
 
         $r->withMinHops(12);
@@ -491,128 +602,11 @@ final class RelationshipTest extends TestCase
 
     public function testGetTypes(): void
     {
-        $r = new Relationship(Relationship::DIR_LEFT);
+        $r = new Relationship(Direction::LEFT);
         $this->assertEmpty($r->getTypes());
 
         $r->withTypes(['a', 'b']);
 
         $this->assertSame(['a', 'b'], $r->getTypes());
-    }
-
-    public function provideVariableLengthRelationshipsWithNameData(): array
-    {
-        return [
-            ['b', 0, 100, Relationship::DIR_UNI, '-[b*0..100]-'],
-            ['b', 5, 5, Relationship::DIR_RIGHT, '-[b*5..5]->'],
-            ['a', 10, null, Relationship::DIR_UNI, '-[a*10..]-'],
-            ['a', null, 10, Relationship::DIR_LEFT, '<-[a*..10]-'],
-        ];
-    }
-
-    public function provideVariableLengthRelationshipsWithTypeData(): array
-    {
-        return [
-            ['', 1, 100, Relationship::DIR_LEFT, '<-[*1..100]-'],
-            ['a', 10, null, Relationship::DIR_LEFT, '<-[:a*10..]-'],
-            [':', null, 10, Relationship::DIR_LEFT, '<-[:`:`*..10]-'],
-        ];
-    }
-
-    public function provideVariableLengthRelationshipsWithPropertiesData(): array
-    {
-        return [
-            [[], 10, 100, Relationship::DIR_LEFT, "<-[*10..100]-"],
-            [[new String_('a')], 10, null, Relationship::DIR_LEFT, "<-[*10.. {`0`: 'a'}]-"],
-            [['a' => new String_('b')], null, 10, Relationship::DIR_LEFT, "<-[*..10 {a: 'b'}]-"],
-        ];
-    }
-
-    public function provideVariableLengthRelationshipsWithNameAndTypeAndPropertiesData(): array
-    {
-        return [
-            ['a', 'a', [], 10, 100, Relationship::DIR_LEFT, "<-[a:a*10..100]-"],
-            ['b', 'a', [new String_('a')], null, 10, Relationship::DIR_LEFT, "<-[b:a*..10 {`0`: 'a'}]-"],
-            ['a', 'b', [new String_('a')], 10, 100, Relationship::DIR_LEFT, "<-[a:b*10..100 {`0`: 'a'}]-"],
-            ['a', '', ['a' => new String_('b')], null, 10, Relationship::DIR_LEFT, "<-[a*..10 {a: 'b'}]-"],
-            ['a', ':', ['a' => new String_('b'), new String_('c')], 10, null, Relationship::DIR_LEFT, "<-[a:`:`*10.. {a: 'b', `0`: 'c'}]-"],
-        ];
-    }
-
-    public function provideWithNameData(): array
-    {
-        return [
-            ['a', Relationship::DIR_UNI, '-[a]-'],
-            ['a', Relationship::DIR_LEFT, '<-[a]-'],
-            ['a', Relationship::DIR_RIGHT, '-[a]->'],
-        ];
-    }
-
-    public function provideAddTypeData(): array
-    {
-        return [
-            ['', Relationship::DIR_LEFT, '<--'],
-            ['a', Relationship::DIR_LEFT, '<-[:a]-'],
-            [':', Relationship::DIR_LEFT, '<-[:`:`]-'],
-        ];
-    }
-
-    public function provideWithTypesData(): array
-    {
-        return [
-            [['', 'a'], Relationship::DIR_LEFT, '<-[:a]-'],
-            [['a', 'b'], Relationship::DIR_LEFT, '<-[:a|b]-'],
-            [['', 'a', 'b'], Relationship::DIR_LEFT, '<-[:a|b]-'],
-            [['a', '', 'b'], Relationship::DIR_LEFT, '<-[:a|b]-'],
-            [['a', 'b', ':'], Relationship::DIR_LEFT, '<-[:a|b|`:`]-'],
-        ];
-    }
-
-    public function provideWithPropertiesData(): array
-    {
-        return [
-            [[], Relationship::DIR_LEFT, "<--"],
-            [[new String_('a')], Relationship::DIR_LEFT, "<-[{`0`: 'a'}]-"],
-            [['a' => new String_('b')], Relationship::DIR_LEFT, "<-[{a: 'b'}]-"],
-            [['a' => new String_('b'), new String_('c')], Relationship::DIR_LEFT, "<-[{a: 'b', `0`: 'c'}]-"],
-            [[':' => new Integer(12)], Relationship::DIR_LEFT, "<-[{`:`: 12}]-"],
-            [['a' => 'b', 'c' => 12, 'd' => 12.38], Relationship::DIR_LEFT, "<-[{a: 'b', c: 12, d: 12.38}]-"],
-        ];
-    }
-
-    public function provideWithNameAndTypeData(): array
-    {
-        return [
-            ['a', '', Relationship::DIR_LEFT, '<-[a]-'],
-            ['a', 'b', Relationship::DIR_LEFT, '<-[a:b]-'],
-        ];
-    }
-
-    public function provideWithNameAndPropertiesData(): array
-    {
-        return [
-            ['a', [], Relationship::DIR_LEFT, "<-[a]-"],
-            ['b', [new String_('a')], Relationship::DIR_LEFT, "<-[b {`0`: 'a'}]-"],
-        ];
-    }
-
-    public function provideWithNameAndTypeAndPropertiesData(): array
-    {
-        return [
-            ['a', 'a', [], Relationship::DIR_LEFT, "<-[a:a]-"],
-            ['b', 'a', [new String_('a')], Relationship::DIR_LEFT, "<-[b:a {`0`: 'a'}]-"],
-            ['a', 'b', [new String_('a')], Relationship::DIR_LEFT, "<-[a:b {`0`: 'a'}]-"],
-            ['a', '', ['a' => new String_('b')], Relationship::DIR_LEFT, "<-[a {a: 'b'}]-"],
-            ['a', ':', ['a' => new String_('b'), new String_('c')], Relationship::DIR_LEFT, "<-[a:`:` {a: 'b', `0`: 'c'}]-"],
-        ];
-    }
-
-    public function provideWithMultipleTypesData(): array
-    {
-        return [
-            ['a', [], [], Relationship::DIR_LEFT, "<-[a]-"],
-            ['b', ['a'], [new String_('a')], Relationship::DIR_LEFT, "<-[b:a {`0`: 'a'}]-"],
-            ['a', ['a', 'b', 'c'], [new String_('a')], Relationship::DIR_LEFT, "<-[a:a|b|c {`0`: 'a'}]-"],
-            ['a', ['a', 'b'], [], Relationship::DIR_LEFT, "<-[a:a|b]-"],
-        ];
     }
 }
